requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
idna==2.10
numpy==1.19.4
Pillow==8.0.1
Pygments==2.7.3
Pypubsub==4.0.3
requests==2.25.0
robotframework==3.2.2
robotframework-pythonlibcore==2.1.0
robotframework-ride==1.7.4.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.5.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.2
wxPython==4.0.7.post2
